+++ src/quacc/recipes/dftb/core.py
@@ -34,7 +34,7 @@ def static_job(
         k-point grid to use.
     **calc_kwargs
         Custom kwargs for the calculator that would override the
-        calculator defaults. Set a value to `None` to remove a pre-existing key
+        calculator defaults. Set a value to `quacc.Remove` to remove a pre-existing key
         entirely. For a list of available keys, refer to the
         `ase.calculators.dftb.Dftb` calculator.
 
@@ -47,9 +47,10 @@ def static_job(
     calc_defaults = {
         "Hamiltonian_": "xTB" if "xtb" in method.lower() else "DFTB",
         "Hamiltonian_MaxSccIterations": 200,
-        "Hamiltonian_Method": method if "xtb" in method.lower() else None,
         "kpts": kpts or ((1, 1, 1) if atoms.pbc.any() else None),
     }
+    if "xtb" in method.lower():
+        calc_defaults["Hamiltonian_Method"] = method
 
     return base_fn(
         atoms,
@@ -83,7 +84,7 @@ def relax_job(
         positions.
     **calc_kwargs
         Custom kwargs for the calculator that would override the
-        calculator defaults. Set a value to `None` to remove a pre-existing key
+        calculator defaults. Set a value to `quacc.Remove` to remove a pre-existing key
         entirely. For a list of available keys, refer to the
         `ase.calculators.dftb.Dftb` calculator.
 
@@ -100,9 +101,10 @@ def relax_job(
         "Driver_MaxSteps": 2000,
         "Hamiltonian_": "xTB" if "xtb" in method.lower() else "DFTB",
         "Hamiltonian_MaxSccIterations": 200,
-        "Hamiltonian_Method": method if "xtb" in method.lower() else None,
         "kpts": kpts or ((1, 1, 1) if atoms.pbc.any() else None),
     }
+    if "xtb" in method.lower():
+        calc_defaults["Hamiltonian_Method"] = method
 
     return base_fn(
         atoms,