+++ src/quacc/recipes/vasp/core.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from quacc import job
+from quacc import job, strip_decorator
 from quacc.recipes.vasp._base import base_fn
 
 if TYPE_CHECKING:
@@ -35,7 +35,7 @@ def static_job(
         File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Custom kwargs for the Vasp calculator. Set a value to
-        `None` to remove a pre-existing key entirely. For a list of available
+        `quacc.Remove` to remove a pre-existing key entirely. For a list of available
         keys, refer to the [quacc.calculators.vasp.vasp.Vasp][] calculator.
 
     Returns
@@ -87,7 +87,7 @@ def relax_job(
         File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Custom kwargs for the Vasp calculator. Set a value to
-        `None` to remove a pre-existing key entirely. For a list of available
+        `quacc.Remove` to remove a pre-existing key entirely. For a list of available
         keys, refer to the [quacc.calculators.vasp.vasp.Vasp][] calculator.
 
     Returns
@@ -158,12 +158,12 @@ def double_relax_job(
     relax2_kwargs = relax2_kwargs or {}
 
     # Run first relaxation
-    summary1 = relax_job.__wrapped__(
+    summary1 = strip_decorator(relax_job)(
         atoms, preset=preset, relax_cell=relax_cell, **relax1_kwargs
     )
 
     # Run second relaxation
-    summary2 = relax_job.__wrapped__(
+    summary2 = strip_decorator(relax_job)(
         summary1["atoms"],
         preset=preset,
         relax_cell=relax_cell,